+++ modules/AWSPowerShell/Cmdlets/WAFV2/AWS.Tools.WAFV2.Completers.psm1
@@ -94,6 +94,17 @@ $WAF2_Completers = {
             break
         }
 
+        # Amazon.WAFV2.Platform
+        {
+            ($_ -eq "Get-WAF2MobileSdkRelease/Platform") -Or
+            ($_ -eq "Get-WAF2MobileSdkReleaseList/Platform") -Or
+            ($_ -eq "New-WAF2MobileSdkReleaseUrl/Platform")
+        }
+        {
+            $v = "ANDROID","IOS"
+            break
+        }
+
         # Amazon.WAFV2.ResourceType
         "Get-WAF2ResourcesForWebACLList/ResourceType"
         {
@@ -151,6 +162,7 @@ $WAF2_Completers = {
 $WAF2_map = @{
     "IPAddressVersion"=@("New-WAF2IPSet")
     "LoggingConfiguration_LoggingFilter_DefaultBehavior"=@("Write-WAF2LoggingConfiguration")
+    "Platform"=@("Get-WAF2MobileSdkRelease","Get-WAF2MobileSdkReleaseList","New-WAF2MobileSdkReleaseUrl")
     "ResourceType"=@("Get-WAF2ResourcesForWebACLList")
     "Scope"=@("Get-WAF2AvailableManagedRuleGroupList","Get-WAF2AvailableManagedRuleGroupVersionList","Get-WAF2IPSet","Get-WAF2IPSetList","Get-WAF2LoggingConfigurationList","Get-WAF2ManagedRuleGroup","Get-WAF2ManagedRuleSet","Get-WAF2ManagedRuleSetList","Get-WAF2RateBasedStatementManagedKey","Get-WAF2RegexPatternSet","Get-WAF2RegexPatternSetList","Get-WAF2RuleGroup","Get-WAF2RuleGroupList","Get-WAF2SampledRequest","Get-WAF2WebACL","Get-WAF2WebACLsList","New-WAF2IPSet","New-WAF2RegexPatternSet","New-WAF2RuleGroup","New-WAF2WebACL","Remove-WAF2IPSet","Remove-WAF2RegexPatternSet","Remove-WAF2RuleGroup","Remove-WAF2WebACL","Test-WAF2Capacity","Update-WAF2IPSet","Update-WAF2ManagedRuleSetVersionExpiryDate","Update-WAF2RegexPatternSet","Update-WAF2RuleGroup","Update-WAF2WebACL","Write-WAF2ManagedRuleSetVersion")
 }
@@ -220,9 +232,11 @@ $WAF2_SelectMap = @{
                "Remove-WAF2WebACL",
                "Get-WAF2ManagedRuleGroup",
                "Remove-WAF2WebACLFromResource",
+               "New-WAF2MobileSdkReleaseUrl",
                "Get-WAF2IPSet",
                "Get-WAF2LoggingConfiguration",
                "Get-WAF2ManagedRuleSet",
+               "Get-WAF2MobileSdkRelease",
                "Get-WAF2PermissionPolicy",
                "Get-WAF2RateBasedStatementManagedKey",
                "Get-WAF2RegexPatternSet",
@@ -235,6 +249,7 @@ $WAF2_SelectMap = @{
                "Get-WAF2IPSetList",
                "Get-WAF2LoggingConfigurationList",
                "Get-WAF2ManagedRuleSetList",
+               "Get-WAF2MobileSdkReleaseList",
                "Get-WAF2RegexPatternSetList",
                "Get-WAF2ResourcesForWebACLList",
                "Get-WAF2RuleGroupList",