+++ modules/AWSPowerShell/Cmdlets/WAFV2/Basic/Get-WAF2MobileSdkReleaseList-Cmdlet.cs
@@ -1,0 +1,261 @@
+/*******************************************************************************
+ *  Copyright 2012-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.WAFV2;
+using Amazon.WAFV2.Model;
+
+namespace Amazon.PowerShell.Cmdlets.WAF2
+{
+    /// <summary>
+    /// Retrieves a list of the available releases for the mobile SDK and the specified device
+    /// platform. 
+    /// 
+    ///  
+    /// <para>
+    /// The mobile SDK is not generally available. Customers who have access to the mobile
+    /// SDK can use it to establish and manage Security Token Service (STS) security tokens
+    /// for use in HTTP(S) requests from a mobile device to WAF. 
+    /// </para><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
+    /// </summary>
+    [Cmdlet("Get", "WAF2MobileSdkReleaseList")]
+    [OutputType("Amazon.WAFV2.Model.ReleaseSummary")]
+    [AWSCmdlet("Calls the AWS WAF V2 ListMobileSdkReleases API operation.", Operation = new[] {"ListMobileSdkReleases"}, SelectReturnType = typeof(Amazon.WAFV2.Model.ListMobileSdkReleasesResponse))]
+    [AWSCmdletOutput("Amazon.WAFV2.Model.ReleaseSummary or Amazon.WAFV2.Model.ListMobileSdkReleasesResponse",
+        "This cmdlet returns a collection of Amazon.WAFV2.Model.ReleaseSummary objects.",
+        "The service call response (type Amazon.WAFV2.Model.ListMobileSdkReleasesResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class GetWAF2MobileSdkReleaseListCmdlet : AmazonWAFV2ClientCmdlet, IExecutor
+    {
+        
+        #region Parameter Platform
+        /// <summary>
+        /// <para>
+        /// <para>The device platform to retrieve the list for.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        [AWSConstantClassSource("Amazon.WAFV2.Platform")]
+        public Amazon.WAFV2.Platform Platform { get; set; }
+        #endregion
+        
+        #region Parameter Limit
+        /// <summary>
+        /// <para>
+        /// <para>The maximum number of objects that you want WAF to return for this request. If more
+        /// objects are available, in the response, WAF provides a <code>NextMarker</code> value
+        /// that you can use in a subsequent call to get the next batch of objects.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Int32? Limit { get; set; }
+        #endregion
+        
+        #region Parameter NextMarker
+        /// <summary>
+        /// <para>
+        /// <para>When you request a list of objects with a <code>Limit</code> setting, if the number
+        /// of objects that are still available for retrieval exceeds the limit, WAF returns a
+        /// <code>NextMarker</code> value in the response. To retrieve the next batch of objects,
+        /// provide the marker from the prior call in your next request.</para>
+        /// </para>
+        /// <para>
+        /// <br/><b>Note:</b> This parameter is only used if you are manually controlling output pagination of the service API call.
+        /// <br/>In order to manually control output pagination, use '-NextMarker $null' for the first call and '-NextMarker $AWSHistory.LastServiceResponse.NextMarker' for subsequent calls.
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("NextToken")]
+        public System.String NextMarker { get; set; }
+        #endregion
+        
+        #region Parameter Select
+        /// <summary>
+        /// Use the -Select parameter to control the cmdlet output. The default value is 'ReleaseSummaries'.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.WAFV2.Model.ListMobileSdkReleasesResponse).
+        /// Specifying the name of a property of type Amazon.WAFV2.Model.ListMobileSdkReleasesResponse will result in that property being returned.
+        /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public string Select { get; set; } = "ReleaseSummaries";
+        #endregion
+        
+        #region Parameter NoAutoIteration
+        /// <summary>
+        /// By default the cmdlet will auto-iterate and retrieve all results to the pipeline by performing multiple
+        /// service calls. If set, the cmdlet will retrieve only the next 'page' of results using the value of NextMarker
+        /// as the start point.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public SwitchParameter NoAutoIteration { get; set; }
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext();
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            if (ParameterWasBound(nameof(this.Select)))
+            {
+                context.Select = CreateSelectDelegate<Amazon.WAFV2.Model.ListMobileSdkReleasesResponse, GetWAF2MobileSdkReleaseListCmdlet>(Select) ??
+                    throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
+            }
+            context.Limit = this.Limit;
+            context.NextMarker = this.NextMarker;
+            context.Platform = this.Platform;
+            #if MODULAR
+            if (this.Platform == null && ParameterWasBound(nameof(this.Platform)))
+            {
+                WriteWarning("You are passing $null as a value for parameter Platform which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            var useParameterSelect = this.Select.StartsWith("^");
+            
+            // create request and set iteration invariants
+            var request = new Amazon.WAFV2.Model.ListMobileSdkReleasesRequest();
+            
+            if (cmdletContext.Limit != null)
+            {
+                request.Limit = cmdletContext.Limit.Value;
+            }
+            if (cmdletContext.Platform != null)
+            {
+                request.Platform = cmdletContext.Platform;
+            }
+            
+            // Initialize loop variant and commence piping
+            var _nextToken = cmdletContext.NextMarker;
+            var _userControllingPaging = this.NoAutoIteration.IsPresent || ParameterWasBound(nameof(this.NextMarker));
+            
+            var client = Client ?? CreateClient(_CurrentCredentials, _RegionEndpoint);
+            do
+            {
+                request.NextMarker = _nextToken;
+                
+                CmdletOutput output;
+                
+                try
+                {
+                    
+                    var response = CallAWSServiceOperation(client, request);
+                    
+                    object pipelineOutput = null;
+                    if (!useParameterSelect)
+                    {
+                        pipelineOutput = cmdletContext.Select(response, this);
+                    }
+                    output = new CmdletOutput
+                    {
+                        PipelineOutput = pipelineOutput,
+                        ServiceResponse = response
+                    };
+                    
+                    _nextToken = response.NextMarker;
+                }
+                catch (Exception e)
+                {
+                    output = new CmdletOutput { ErrorResponse = e };
+                }
+                
+                ProcessOutput(output);
+                
+            } while (!_userControllingPaging && AutoIterationHelpers.HasValue(_nextToken));
+            
+            if (useParameterSelect)
+            {
+                WriteObject(cmdletContext.Select(null, this));
+            }
+            
+            
+            return null;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.WAFV2.Model.ListMobileSdkReleasesResponse CallAWSServiceOperation(IAmazonWAFV2 client, Amazon.WAFV2.Model.ListMobileSdkReleasesRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS WAF V2", "ListMobileSdkReleases");
+            try
+            {
+                #if DESKTOP
+                return client.ListMobileSdkReleases(request);
+                #elif CORECLR
+                return client.ListMobileSdkReleasesAsync(request).GetAwaiter().GetResult();
+                #else
+                        #error "Unknown build edition"
+                #endif
+            }
+            catch (AmazonServiceException exc)
+            {
+                var webException = exc.InnerException as System.Net.WebException;
+                if (webException != null)
+                {
+                    throw new Exception(Utils.Common.FormatNameResolutionFailureMessage(client.Config, webException.Message), webException);
+                }
+                throw;
+            }
+        }
+        
+        #endregion
+        
+        internal partial class CmdletContext : ExecutorContext
+        {
+            public System.Int32? Limit { get; set; }
+            public System.String NextMarker { get; set; }
+            public Amazon.WAFV2.Platform Platform { get; set; }
+            public System.Func<Amazon.WAFV2.Model.ListMobileSdkReleasesResponse, GetWAF2MobileSdkReleaseListCmdlet, object> Select { get; set; } =
+                (response, cmdlet) => response.ReleaseSummaries;
+        }
+        
+    }
+}